+++ login.html
@@ -58,5 +58,6 @@
       </p>
     </div>
   </body>
-  <script src="js/auth.js" defer></script>
+  <script type="module" src="js/validationUtils.js" defer></script>
+  <script type="module" src="js/login.js" defer></script>
 </html>
