+++ signup.html
@@ -81,5 +81,6 @@
       </p>
     </div>
   </body>
-  <script src="js/auth.js" defer></script>
+  <script type="module" src="js/validationUtils.js" defer></script>
+  <script type="module" src="js/signup.js" defer></script>
 </html>
